+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ### Added
 
-- [#2166](https://github.com/slimphp/Slim/pull/2166) Added Middleware\OutputBuffering
+- [#2166](https://github.com/slimphp/Slim/pull/2166) Added Middleware\OutputBufferingMiddleware
 
 ### Deprecated